updating changes from Kryha-repo

+++ packages/noir/scripts/export_sol_verifiers.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if ! command -v nargo &> /dev/null
+then
+    echo "nargo could not be found, exiting..."
+    echo ""
+    exit 2
+fi
+
 echo "exporting sol verifiers..."
 rm ../hardhat/contracts/verifiers/*
 for d in circuits/*/ ; do
